# HomeCompassApi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "HomeCompassApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/Facility": {
      "post": {
        "tags": [
          "Facility"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Facility"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Facility"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ReadFacilitiesDTO"
              }
            }
          }
        }
      }
    },
    "/Facility/{id}": {
      "get": {
        "tags": [
          "Facility"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Facility"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Facility"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateFacilityDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Facility"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Facility/contributor/{id}": {
      "get": {
        "tags": [
          "Facility"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Facility"
              }
            }
          }
        }
      }
    },
    "/Facility/bycategory/{categoryId}": {
      "get": {
        "tags": [
          "Facility"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Facility"
              }
            }
          }
        }
      }
    },
    "/Facility/page": {
      "post": {
        "tags": [
          "Facility"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PageDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ReadFacilitiesDTO"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Facility": {
      "required": [
        "location",
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "minLength": 1,
          "type": "string"
        },
        "location": {
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "target": {
          "type": "string"
        },
        "resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Resource"
          }
        },
        "days": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "hours": {
          "format": "int32",
          "type": "integer"
        },
        "photoUrl": {
          "format": "uri",
          "type": "string"
        },
        "contactInformaton": {
          "format": "tel",
          "type": "string"
        },
        "contributorId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "PageDTO": {
      "required": [
        "index",
        "size"
      ],
      "type": "object",
      "properties": {
        "index": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 1,
          "type": "integer"
        },
        "size": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 1,
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "ReadFacilitiesDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "contactInformaton": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Resource"
          }
        },
        "target": {
          "type": "string"
        },
        "photoUrl": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Resource": {
      "required": [
        "isAvailable",
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "minLength": 1,
          "type": "string"
        },
        "isAvailable": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateFacilityDTO": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "minLength": 1,
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "target": {
          "type": "string"
        },
        "contactInformation": {
          "type": "string"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "hours": {
          "format": "int32",
          "type": "integer"
        },
        "resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Resource"
          }
        },
        "days": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "photoUrl": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}